mysite/requirements.txt
boto==2.42.0
configparser==3.5.0
Django==1.10.1
django-autofixture==0.12.1
django-bootstrap3==7.0.1
django-filter==0.14.0
django-formtools==1.0
django-storages==1.5.1
djangorestframework==3.4.6
Markdown==2.6.6
flake8==3.0.4
invoke==0.13.0
mccabe==0.5.2
Pillow==3.3.1
pycodestyle==2.0.0
pyflakes==1.2.3
